Arghl. Again a stupid bug caused by float having poor precision. Reverting r700026 and changing floats to doubles again. I'd probably like to change even the ones interfacing with OpenGL which I've left for now.
svn path=/trunk/KDE/kdebase/workspace/; revision=707987

+++ deleted.h
@@ -35,7 +35,7 @@ class Deleted
         void copyToDeleted( Toplevel* c );
         virtual ~Deleted(); // deleted only using unrefWindow()
         int delete_refcount;
-        float window_opacity;
+        double window_opacity;
         int desk;
         QRect contentsRect; // for clientPos()/clientSize()
     };
