+++ effects/desktopgrid.h
@@ -45,7 +45,7 @@ class DesktopGridEffect
         void paintSlide( int mask, QRegion region, const ScreenPaintData& data );
         void paintScreenDesktop( int desktop, int mask, QRegion region, ScreenPaintData data );
         void slideDesktopChanged( int old );
-        float progress;
+        double progress;
         bool activated;
         int painting_desktop;
         int hover_desktop;
