+++ effects/shadow.cpp
@@ -25,7 +25,7 @@ ShadowEffect::ShadowEffect()
     KConfigGroup conf = effects->effectConfig("Shadow");
     shadowXOffset = conf.readEntry( "XOffset", 10 );
     shadowYOffset = conf.readEntry( "YOffset", 10 );
-    shadowOpacity = (float)conf.readEntry( "Opacity", 0.2 );
+    shadowOpacity = conf.readEntry( "Opacity", 0.2 );
     shadowFuzzyness = conf.readEntry( "Fuzzyness", 10 );
 
     QString shadowtexture =  KGlobal::dirs()->findResource("data", "kwin/shadow-texture.png");
@@ -125,7 +125,7 @@ void ShadowEffect::drawShadow( EffectWindow* window, int mask, QRegion region, W
     mShadowTexture->bind();
     // Take the transparency settings and window's transparency into account.
     // Also make the shadow more transparent if we've made it bigger
-    glColor4f(0, 0, 0, shadowOpacity * data.opacity * (window->width() / (float)w) * (window->height() / (float)h));
+    glColor4f(0, 0, 0, shadowOpacity * data.opacity * (window->width() / (double)w) * (window->height() / (double)h));
     glTexEnvi(GL_TEXTURE_ENV, GL_TEXTURE_ENV_MODE, GL_MODULATE);
     // We have two elements per vertex in the verts array
     int verticesCount = verts.count() / 2;
